Adding memory size and CUDA version checker

+++ commit/cudaoperator.pyx
@@ -98,7 +98,7 @@ cdef class CudaLinearOperator :
         self.n2 = self.nR*self.nF + self.nT*self.nE + self.nI*self.nV
 
         
-        cdef float gpumem = 1E-6 * (28*self.n + 6*self.nE + 8*(self.nF) + 16*self.nV + 4*(self.nR*self.ndirs*self.nS + self.nT*self.ndirs*self.nS + self.nI*self.nS + self.n1 + self.n2) )
+        cdef double gpumem = 1E-6 * (28.0*self.n + 6.0*self.nE + 8.0*(self.nF) + 16.0*self.nV + 4.0*(self.nR*self.ndirs*self.nS + self.nT*self.ndirs*self.nS + self.nI*self.nS + self.n1 + self.n2) )
         print('Required GPU Memory = %f MB' % gpumem)
         if gpumem > 8000.0:
             raise RuntimeError( 'GPU Memory exceeded!!!!!!' )
